requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
attrs==24.2.0
bitsandbytes==0.42.0
certifi==2024.8.30
charset-normalizer==3.3.2
datasets==3.0.1
dill==0.3.8
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
psutil==6.1.0
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
sympy==1.13.3
tokenizers==0.20.0
torch==2.2.2
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.14.0
